delete Song

+++ src/controllers/SongController/songController.js
@@ -198,7 +198,6 @@ const findUnverified = async(req, res) => {
 const verifiedSong = async(req, res) => {
     try {
         const { Name } = req.body;
-
         if (!Name) return res.status(404).json({ error: "Missing song name"});
 
         const unverifiedSong = await Song.findOne({ Name: Name }).exec();
@@ -217,11 +216,14 @@ const verifiedSong = async(req, res) => {
 const deleteSong = async (req, res) => {
     try {
         const { Name } = req.body;
-        const deleteSong = await User.findOne({ Name: Name });
+        if (!Name) return res.status(404).json({ error: "Missing song name"});
+
+        const deleteSong = await Song.findOne({ Name: Name }).exec();
         if (!deleteSong){
-            return res.status(404).json({ message:"User not found" });
+            return res.status(400).json({ message:"Song not found" });
         }
-        const name = deleteSong.name;
+        
+        const name = deleteSong.Name;
         await Song.deleteOne({ Name: Name });
         res.status(200).json({ message: `Song ${name} has been deleted`});
     } catch (error) {
